+++ build_jar.py
@@ -48,6 +48,11 @@ def main():
     project = util.get_project(project_dict[HOME_REGION])
     print("project: {} ({})".format(project.name, project.get_id()))
 
+
+    # Figure out what the current version is
+    version_id = util.get_version_id(top_dir)
+    print("version: {}".format(version_id))
+
     # Set the folder, build one if necessary
     if args.folder is not None:
         folder = args.folder
@@ -65,14 +70,12 @@ def main():
     if args.release:
         multi_region = True
 
-    # Figure out what the current version is
-    version_id = util.get_version_id(top_dir)
-    print("version: {}".format(version_id))
-
     # build the asset
-    home_ad = util.build(project, folder, version_id, top_dir)
+    util.build(project, folder, version_id, top_dir)
 
-    ad_all = [home_ad]
+    # build the compiler jar file
+    jar_path = util.build_compiler_jar(version_id, top_dir, project_dict)
+
     if multi_region:
         # download dxWDL runtime library
         home_rec = dxpy.DXRecord(home_ad.asset_id)
@@ -92,11 +95,6 @@ def main():
                     raise Exception("No project configured for region {}".format(region))
                 dest_proj = util.get_project(proj)
                 dest_ad = util.copy_across_regions(rtlib_path, home_rec, region, dest_proj, folder)
-                ad_all.append(dest_ad)
-
-    # build the final jar file, containing a list of the per-region
-    # assets
-    jar_path = util.build_final_jar(version_id, top_dir, ad_all)
 
     # Upload compiler jar file
     if args.release: